# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tv_shows": {
            "get": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "Get List Data",
                "description": "enter your  here",
                "operationId": "tv_shows",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "tv_shows",
                "description": "tv_shows",
                "operationId": "Tv_Shows",
                "requestBody": {
                    "description": "form tv_shows",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tv_show_name"
                                ],
                                "properties": {
                                    "tv_show_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/tv_shows/{id}": {
            "get": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "tv_shows/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "Delete artist",
                "description": "-",
                "operationId": "tv_shows/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/tv_shows{id}": {
            "put": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "Update artist",
                "description": "-",
                "operationId": "tv_shows/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tv_show_name"
                                ],
                                "properties": {
                                    "tv_show_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/years": {
            "get": {
                "tags": [
                    "Years"
                ],
                "summary": "Get List years Data",
                "description": "enter your years here",
                "operationId": "years",
                "responses": {
                    "default": {
                        "description": "return array model years"
                    }
                }
            },
            "post": {
                "tags": [
                    "Years"
                ],
                "summary": "years",
                "description": "years",
                "operationId": "Years",
                "requestBody": {
                    "description": "form years",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "year"
                                ],
                                "properties": {
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/years/{id}": {
            "get": {
                "tags": [
                    "Years"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "years/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Years"
                ],
                "summary": "Update years",
                "description": "-",
                "operationId": "years/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "year"
                                ],
                                "properties": {
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Years"
                ],
                "summary": "Delete years",
                "description": "-",
                "operationId": "years/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}